pyproject.toml
[project]
name = "claude-telegram-bot"
version = "0.1.0"
description = "Telegram bot for interacting with Claude Code CLI"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
]

[project.scripts]
claude-tg-bot = "claude_telegram_bot.bot:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_telegram_bot"]
